feat: 캘린더 수정, 삭제 시 editRole 검증 추가

+++ src/main/java/com/example/alcdiary/application/impl/CalenderUseCaseImpl.java
@@ -6,6 +6,7 @@ import com.example.alcdiary.application.command.UpdateCalenderCommand;
 import com.example.alcdiary.application.result.FindCalenderResult;
 import com.example.alcdiary.domain.enums.DrinkType;
 import com.example.alcdiary.domain.service.CalenderService;
+import com.example.alcdiary.domain.service.UserCalenderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,26 +15,30 @@ import org.springframework.stereotype.Service;
 public class CalenderUseCaseImpl implements CalenderUseCase {
 
     private final CalenderService calenderService;
+    private final UserCalenderService userCalenderService;
 
     @Override
     public FindCalenderResult find(String userId, Long calenderId) {
+        // TODO: user_calender에서 join 걸어서 찾기
         return new FindCalenderResult().
                 fromModel(calenderService.find(userId, calenderId));
     }
 
     @Override
     public void create(CreateCalenderCommand command) {
-        CreateCalenderCommand.builder().drinkReport(DrinkType.calculate(command.getDrinks()));
-        calenderService.save(command);
+        command.setDrinkReport(DrinkType.calculate(command.getDrinks()));
+        calenderService.saveUserAndCalenderData(command);
     }
 
     @Override
     public void update(UpdateCalenderCommand command, Long calenderId) {
+        if (!userCalenderService.validateUserRole(command.getUserId(), calenderId)) return;
         calenderService.update(command, calenderId);
     }
 
     @Override
     public void delete(Long calenderId, String userId) {
+        if (!userCalenderService.validateUserRole(userId, calenderId)) return;
         calenderService.delete(calenderId, userId);
     }
 }
